Fixed Issue With Object Set Lock
In the original TTS-Score-Sheet-Lock branch that was merged, setting the object data to Locked: true, is actually incorrect as the onLoad() call will attempt to reposition, and if it was locked in the JSON save, then it won't move.

Added self.setLock(true) in the onLoad() function after the self.setPosition() call.