Updated deleteTable() in ExtractTerrain Object
The deleteTable() function was iterating over nil objects. It is unknown if these objects have been removed, their GUIDs have changed, or some other issue.

But, the deleteTerrain() function does need a little bit of robustness in that before it attempts to call obj.destroy(), check if the object is nil.